+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 ##### Enhancements
 
-* None.  
+* Add support for disableMainThreadChecker and stopOnEveryMainThreadCheckerIssue flags
+  [Jacek Suliga](https://github.com/jmkk)
+  [#619](https://github.com/CocoaPods/Xcodeproj/pull/619)
 
 ##### Bug Fixes